requirements.txt
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
